% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_con.R
\name{init_con}
\alias{init_con}
\title{Initialise Serial Connection to PMT}
\usage{
init_con(port)
}
\arguments{
\item{port}{\link{character} (\strong{required}): port name to be used for the connection}
}
\value{
Returns object of class \code{'serialConnection'}.
}
\description{
Establish the serial connect via the RS232 interface to the
PMT using \link[serial:serialConnection]{serial::serialConnection} with preset values.
}
\examples{
#TODO

}
\author{
Sebastian Kreutzer,  Institute of Geography, Heidelberg University (DE)
}
\keyword{internal}
