% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ports.R
\name{list_ports}
\alias{list_ports}
\title{List available ports}
\usage{
list_ports()
}
\value{
\link{character} vector with available ports
}
\description{
A convenience wrapper around \link[serial:listPorts]{serial::listPorts}
}
\examples{
list_ports()

}
\author{
Sebastian Kreutzer,  Institute of Geography, Heidelberg University (DE)
}
\keyword{internal}
