% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LyoluminescenceApp-package.R
\docType{package}
\name{LyoluminescenceApp-package}
\alias{LyoluminescenceApp-package}
\title{LyoluminescenceApp\cr}
\description{
TODO
}
\details{
TODO more content to add
}
\keyword{package}
