% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_count_data.R
\name{read_count_data}
\alias{read_count_data}
\title{Reads PMT count data}
\usage{
read_count_data(con, freq = 1)
}
\arguments{
\item{con}{\link[serial:serialConnection]{serial::serialConnection} (\strong{required}): a \code{serialConnection} object
returned by either \link{init_con} or \link[serial:serialConnection]{serial::serialConnection}}

\item{freq}{\link{integer} (\emph{with default}): readout frequency to control the cycles
of the readout}
}
\value{
\link{numeric} vector with count data
}
\description{
Reads and processes the PMT count data
}
\examples{
#TODO

}
\author{
Sebastian Kreutzer,  Institute of Geography, Heidelberg University (DE)
}
\keyword{internal}
