% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Lyoluminescence App}
\usage{
run_app(browser = TRUE, display = "auto")
}
\arguments{
\item{browser}{\link{logical} (\emph{with default}): enable/disable app start in browser}

\item{display}{\link{character} (\emph{with default}): set display mode of the app \link[shiny:runApp]{shiny::runApp}}
}
\value{
A \link[shiny:shiny-package]{shiny::shiny-package} application object.
}
\description{
Controls the shiny app start
}
\examples{
\dontrun{
run_app()
}

}
\seealso{
\link[shiny:shinyApp]{shiny::shinyAppDir}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
}
\concept{shiny}
