% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_cmd.R
\name{send_cmd}
\alias{send_cmd}
\title{Send Commands to the Hamamatsu PMT}
\usage{
send_cmd(con, cmd)
}
\arguments{
\item{con}{\link[serial:serialConnection]{serial::serialConnection} (\strong{required}): a \code{serialConnection}
object returned by either \link{init_con} or \link[serial:serialConnection]{serial::serialConnection}}

\item{cmd}{\link{character} (\strong{required}): command to send to the PMT following the documentation}
}
\value{
\link{list} object with the connection and the response from the PMT
}
\description{
Send commands to the PMT and formats the input following the documentation
and takes care of the particularities of this PMT.
}
\examples{
#TODO

}
\author{
Sebastian Kreutzer,  Institute of Geography, Heidelberg University (DE)
}
\keyword{internal}
