% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_module_measurement.R
\name{server_module_measurement}
\alias{server_module_measurement}
\title{Server module}
\usage{
server_module_measurement(id, user_settings)
}
\arguments{
\item{id}{module id}

\item{user_settings}{}
}
\description{
Server module
}
\seealso{
Other modules: 
\code{\link{server_module_static}()},
\code{\link{ui_module_measurement}()},
\code{\link{ui_module_static_about}()},
\code{\link{ui_module_static_news}()}
}
\concept{modules}
\keyword{internal}
