% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_module_measurement.R
\name{ui_module_measurement}
\alias{ui_module_measurement}
\title{Measurement UI}
\usage{
ui_module_measurement(id)
}
\arguments{
\item{id}{\link{character} (\strong{required}) vector to be used for the namespace.}
}
\description{
Measurement UI
}
\seealso{
Other modules: 
\code{\link{server_module_measurement}()},
\code{\link{server_module_static}()},
\code{\link{ui_module_static_about}()},
\code{\link{ui_module_static_news}()}
}
\concept{modules}
\keyword{internal}
