% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_module_static_about.R
\name{ui_module_static_about}
\alias{ui_module_static_about}
\title{About UI}
\usage{
ui_module_static_about(id)
}
\arguments{
\item{id}{\link{character} (\strong{required}) vector to be used for the namespace.}
}
\description{
About UI
}
\seealso{
Other modules: 
\code{\link{server_module_measurement}()},
\code{\link{server_module_static}()},
\code{\link{ui_module_measurement}()},
\code{\link{ui_module_static_news}()}
}
\concept{modules}
\keyword{internal}
