% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LyoluminescenceApp-package.R
\docType{package}
\name{LyoluminescenceApp-package}
\alias{LyoluminescenceApp}
\alias{LyoluminescenceApp-package}
\title{LyoluminescenceApp}
\description{
Simple graphical user interface to measure lyoluminescence
with an Hamamatsu PMT.
}
\details{
This application interfaces an Hamamatsu H9319-11 connected
via  RS-232 interface and enables recording signals. Particularly, for
lyoluminescence measurements.
}
\author{
\strong{Maintainer}: Sebastian Kreutzer \email{sebastian.kreutzer@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-0734-2199}{ORCID})

Authors:
\itemize{
  \item Marco Colombo (\href{https://orcid.org/0000-0001-6672-0623}{ORCID})
}

Other contributors:
\itemize{
  \item Magdalena Biernacka (\href{https://orcid.org/0000-0001-6283-0550}{ORCID}) [contributor]
}

}
\keyword{package}
